LayoutTests/websocket/tests/handshake-error-expected.txt

CONSOLE MESSAGE: line 0: No response code found: ThisWillCauseHandshakeError
Handshake error test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Closed.
PASS ws.readyState is 2
PASS successfullyParsed is true

TEST COMPLETE
